Change default wrapper tag to generate valid html output on collections
Previously we were generating a `p` wrapper with an `ul` list inside,
which is invalid html. Changing the default wrapper to a div should be
harmless in most scenarios, since people should be relying on the css
class to style and both are block elements, so we should be good.

In case someone is relying on the `p` wrapper, it's possible to get this
functionality back on their app simply by setting the `wrapper_tag` to
generate a `:p` tag again.

This fixes 12 failing tests out of 14 on Rails 4.2, and it was caught
due to the changes in Rails 4.2 HTML sanitization which now uses
nokogiri to validate the generate HTML on assert_select.

More information can be found on the release notes:
http://guides.rubyonrails.org/4_2_release_notes.html#assert-select