Fix last 2 failures on Rails 4.2
These are due to the way assert_select works now on Rails 4.2, where the
output must be tested without escaping, even though the HTML is actually
escaped.

We need to maintain a conditional to test with both 4.2+ and previous
versions, since the assert_select method works differently, even though
the output is the same in all versions.

More information can be found in the release notes:
http://guides.rubyonrails.org/4_2_release_notes.html#assert-select