Respect `nil` value when provided, by skipping model attribute value
Don't try to get attribute from object if value option is specified but
is `nil`. e.g.:

    <%= f.attribute :card_no, value: @card_no %>

If `@card_no` returned `nil`, we'd fallback to fetching the `card_no`
attribute value from the model passed to `show_for`. This skips it,
effectively considering it as "not specified".

If you need the previous behavior, you'll have to conditionally do it
yourself:

    <%= f.attribute :card_no, value: @card_no || obj.card_no %>