Run Rails head tests with whatever rack version they require
Don't lock on rack/rack main otherwise it can cause unexpected failures
that even rails/rails is not aware of. The show_for build with Rails
main is currently failing and not bundling against rack main should fix
it.

Build failure:

    /home/runner/work/show_for/show_for/vendor/bundle/ruby/2.7.0/bundler/gems/rails-10c0b5939f8d/actionpack/lib/action_dispatch/http/request.rb:27:in
    `<class:Request>': uninitialized constant Rack::Request::Env (NameError)

Change in `rack` main that removed the `Env` module:
https://github.com/rack/rack/commit/d0c6efc666ede26768f33935f00530629690369a